Move stats components and error display
